Refactor: RP-161 heat가 double로 반환되도록

+++ src/main/java/com/cherrypick/backend/domain/deal/application/service/DealService.java
@@ -324,7 +324,7 @@ public class DealService {
                     user != null ? user.getNickname() : null,
                     user != null ? user.getBadge().getBadgeId() : null,
                     deal.getCreatedAt().toString(),
-                    (int) deal.getHeat(),
+                    deal.getHeat(),
                     (int) likeCount,
                     (int) commentCount,
                     deal.isSoldOut()
