+++ src/main/java/com/cherrypick/backend/domain/deal/domain/service/DealSearchResponseFactory.java
@@ -79,7 +79,7 @@ public class DealSearchResponseFactory
                     user != null ? user.getNickname() : null,
                     user != null ? user.getBadge().getBadgeId() : null,
                     deal.getCreatedAt().toString(),
-                    (int) deal.getHeat(),
+                    deal.getHeat(),
                     (int) likeCount,
                     (int) commentCount,
                     deal.isSoldOut()
